移除四模式外部路径的过度设计
以后再恢复
